requirements.txt
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
bcrypt==4.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.2
debugpy==1.6.6
decorator==5.1.1
executing==1.2.0
flake8==6.0.0
idna==3.4
iniconfig==2.0.0
ipykernel==6.22.0
ipython==8.11.0
isort==5.12.0
jedi==0.18.2
jupyter_client==8.1.0
jupyter_core==5.3.0
markdown-it-py==3.0.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
nest-asyncio==1.5.6
packaging==23.1
parso==0.8.3
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.1.1
pluggy==1.2.0
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.15.1
PyJWT==2.8.0
PyQt6==6.4.2
pyqt6-plugins==6.4.2.2.3
PyQt6-Qt6==6.4.3
PyQt6-sip==13.5.1
pyqt6-tools==6.4.2.3.3
PySide6==6.5.0
PySide6-Addons==6.5.0
PySide6-Essentials==6.5.0
pyTelegramBotAPI==4.10.0
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pyzmq==25.0.2
qt6-applications==6.4.3.2.3
qt6-tools==6.4.3.1.3
requests==2.28.2
rich==13.4.2
shiboken6==6.5.0
six==1.16.0
stack-data==0.6.2
tornado==6.2
traitlets==5.9.0
urllib3==1.26.15
wcwidth==0.2.6
